split up script, make sure descriptions are available, use scrollx instead of responsive extension
